Move library item identification logic into own module
